fix: mobx removed, added exchangersList, switcher to be (now basic)
