+++ src/components/exchangersList/styles.module.scss
@@ -1,0 +1,12 @@
+.exchangers__body {
+  .exchangers__title {
+    font-size: 14px;
+    font-weight: 500;
+    text-align: center;
+    border-bottom: 0.5px solid rgba(0, 0, 0, 0.2);
+    padding-bottom: 15px;
+  }
+  .exchangers__cards {
+    padding: 15px 0px;
+  }
+}
