+++ src/components/main/styles.module.scss
@@ -46,17 +46,5 @@
       cursor: pointer;
       margin-bottom: 25px;
     }
-    &__body {
-    }
-    &__title {
-      font-size: 14px;
-      font-weight: 500;
-      text-align: center;
-      border-bottom: 0.5px solid rgba(0, 0, 0, 0.2);
-      padding-bottom: 15px;
-    }
-    &__cards {
-      padding: 15px 0px;
-    }
   }
 }
